revert to a fully embedded API
